+++ Countly.m
@@ -443,6 +443,14 @@ NSString* previousEventID;
 #endif
 }
 
+- (void)changeDeviceIDWithMerge:(NSString * _Nullable)deviceID {
+    [self setNewDeviceID:deviceID onServer:YES];
+}
+
+- (void)changeDeviceIDWithoutMerge:(NSString * _Nullable)deviceID {
+    [self setNewDeviceID:deviceID onServer:NO];
+}
+
 - (void)setNewDeviceID:(NSString *)deviceID onServer:(BOOL)onServer
 {
     CLY_LOG_I(@"%s %@ %d", __FUNCTION__, deviceID, onServer);